requirements.txt
pika==1.3.2
pika-stubs==0.1.3
psycopg2-binary==2.9.9
pytest==8.2.2
python-binance==1.0.19
python-dotenv==1.0.1
typing_extensions==4.12.2